Crate ink_stroke_modeler_rs

Re-exports

pub use error::ModelerError;

Modules

error

Structs

ModelerInput

struct holding all information for input event

ModelerParams

all parameters for the modeler

ModelerResult

result struct contains the position, time, presusre as well as the velocity and acceleration data

StrokeModeler

This class models a stroke from a raw input stream. The modeling is performed in several stages

Enums

ModelerInputEventType

modeler Input event Type